Make --select flag offer all passports.
This allows for cases where matches need to be overridden.

The original command was largely only useful when you updated the passport
definition or accidentally mangled the user entry in the repo git config and
needed to reset it. I very, very rarely needed to do either of those things, but
I often ran into situations where I needed to one-off pick a different passport
than the usual.